Log: 2nd game research + ref board
Hi @pippinbarr! I've decided on a theme for the next game
(cryptocurrency mining) and have a few games that'll experiment with.
I'll be back in Montréal this Wednesday and I'm doing a physical
prototype work session this Friday morning. I think we could get
together to discuss the results Friday afternoon, around 4-5pm. Would
that be good for you? Cheers!