Logs for GambiCon03 first work session.
Hi @pippinbarr! Here are my notes and reflections about the current
stage of the third game controller project. I'm planning to do some more
work on it later this week (Friday probably), with the goal to have a
playable version of the controller after the next work session. Let me
know what you think of it. Thanks! :)
Cheers!